+++ hubdle-version-catalog/CHANGELOG.md
@@ -14,9 +14,13 @@
 
 ### Updated
 
+- `com.vanniktech.maven.publish:com.vanniktech.maven.publish.gradle.plugin -> 0.36.0`
 - `gradle -> 9.3.0`
 - `com.gradle.develocity:com.gradle.develocity.gradle.plugin -> 4.3.1`
 - `androidx.compose.animation:animation-core -> 1.10.1`
+- `com.android.library:com.android.library.gradle.plugin -> 9.0.0`
+- `com.android.application:com.android.application.gradle.plugin -> 9.0.0`
+- `com.android.tools.build:gradle -> 9.0.0`
 - `androidx.compose.ui:ui-util -> 1.10.1`
 - `androidx.compose.ui:ui-tooling-preview -> 1.10.1`
 - `androidx.compose.ui:ui-tooling -> 1.10.1`